Remove non-functional if-failed code and warn
The way that the `:if-failed` argument to `with-ftp-connection` is
handled does not work. `make-instance` should never return nil, and
any errors signalled within `initialize-instance` will still be
signalled. I feel it would probably be futile to try and make it work
for the errors which this library signals, as `connect-to-server`
calls other functions which don't have an argument to avoid
signalling, so the client may be better served with `ignore-errors` or
similar.